refactor: separate waiting for a lock from turning a file into a table
The retry that runs when another connection holds the lock a load needs --
its two budgets, the backoff, and the two functions that decide whether an
error is a lock at all -- is a concern of its own, and it sat in the middle
of the file about reading inputs into tables.

stream_processor.go drops from 1240 lines to 1127. The tests stay in
concurrency_test.go, where they belong: each one runs concurrent loads, and
that is the file about what happens when loads collide.
